requirements.txt
asgiref==3.5.2
Django==4.1.4
django-ipware==4.0.2
gunicorn==20.1.0
mysqlclient==2.1.1
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.2.0
